Fix imports for shared coding helpers
